01-api_assincronas_fastApi/fastApi_blog/requirements.txt
# JWT para autenticação
PyJWT
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy
asyncpg
alembic
psycopg2-binary
